UPBGE: Refactor material attributes management.
The attributes management inherited from the BFBGE conception, these attributes
were computed in blender shader, material or custom shader and set globally
through the rasterizer with calls like SetAttribs/SetTexCo. Then these
attributes were pulled back in VBO storage when creating the VAO.
The issues of this conception is the useless process of updating global
attributes as the VBO storage need it only once at its creation. Also the VBO
storage is used for both display array storage (vertices and indices) and for
attributes storage (VAOs per draw type), so when reconstructing the attributes
was needed the VBO was also destructed.

To avoid these issues first of all the VBO and VAO are split in two different
class, respectively RAS_DisplayArrayStorage and RAS_AttributeArrayStorage.
RAS_DisplayArrayStorage is owned by RAS_DisplayArray and is dedicated to wrap
RAS_StorageVbo following a pimpl idiom. This storage is created via function
RAS_IDisplayArray::ConstructStorage called from RAS_DisplayArrayBucket the first
render time when the modified flag STORAGE_INVALID is set, this is designed to
delay storage creation to handle asynchronous libloading where the OpenGL
context is inaccessible. Excepted these the functions are IndexPrimitves[...]...
do the same work as before.
RAS_AttributeArrayStorage is owned by a new class named RAS_AttributeArray, and
this last class is owned by the display array bucket using it. The purpose of
RAS_AttributeArray is to store the list of attributes description (location,
type, layer, flag) and lazily create a RAS_DisplayArrayStorage depending on the
drawing mode used. This process is done by the function GetStorage called at
every render step in RAS_DisplayArrayBucket::UpdateActiveMeshSlots. Similar to
RAS_DisplayArrayStorage, RAS_AttributeArrayStorage wrap RAS_StorageVao using
pimpl and offers the function Bind/UnbindPrimitives. RAS_DisplayArrayStorage and
RAS_AttributeArrayStorage are both referenced in RAS_DisplayArrayNodeData to be
used by node callbacks from display array bucket level and mesh slot level.

As described before the attributes are not anymore a global variable, instead
they are build and get in the same time RAS_AttributeArray is created via a call
to RAS_IPolyMaterial::GetAttribs which redirect (at KX_BlenderMaterial level) to
functions from BL_BLenderShader or BL_Shader.
In consideration BL_BLenderShader remove its parsing and store mechanism. The
display array bucket is noticed that the attributes changed by checking a update
client (CM_UpdateClient) tracked to the used material, to do so
RAS_IPolyMaterial is inherited from CM_UpdateServer and use two flags:
ATTRIBUTES_MODIFIED and SHADER_MODIFIED. These both flags are viewed from the
display array bucket and set by BL_Shader and BL_BLenderShader.

Because the display array storage is created in the constructor of the display
array it's not possible to call any virtual functions from the display array,
the previous function used for memory format of the used vertex data type can't
be used anymore, to fill this gap, a struct named RAS_VertexDataMemoryFormat is
implemented and initialized from the function
RAS_VertexData<>::GetMemoryFormat(), this struct is now used in both
RAS_DisplayArrayStorage and RAS_AttributeArrayStorage.

Previously the VAO was optional because instancing was poorly mixed with the VAO
behavior. The correct behavior is while the VAO is bound to bind once the
instancing buffer, bind the attributes and unbind the buffer. By doing this
there's no conflict of attributes and buffer. In consideration all unbind
instancing function are now unused and so removed.

+++ source/gameengine/Rasterizer/RAS_Vertex.h
@@ -36,8 +36,8 @@
 /// Struct used to pass the vertex format to functions.
 struct RAS_VertexFormat
 {
-	unsigned short uvSize;
-	unsigned short colorSize;
+	uint8_t uvSize;
+	uint8_t colorSize;
 };
 
 /// Operators used to compare the contents (uv size, color size, ...) of two vertex formats.
@@ -54,7 +54,7 @@ public:
 private:
 	unsigned int m_origindex;
 	short m_softBodyIndex;
-	short m_flag;
+	uint8_t m_flag;
 
 public:
 	RAS_VertexInfo(unsigned int origindex, bool flat);
@@ -75,12 +75,12 @@ public:
 		m_softBodyIndex = sbIndex;
 	}
 
-	inline const short GetFlag() const
+	inline const uint8_t GetFlag() const
 	{
 		return m_flag;
 	}
 
-	inline void SetFlag(const short flag)
+	inline void SetFlag(const uint8_t flag)
 	{
 		m_flag = flag;
 	}
